requirements.txt
py-algorand-sdk==1.5.0
pyteal==0.7.0
PyYAML==5.4.1
